feat: sum p and d components first
This prepares for increasing the size of the integrator, which in turn,
would result in unnecessary synthesized resources in actual hardware
using the previous summing strategy. By summing P and D first, then
adding in I in the next pipelined stage, the first adder is smaller if
the integrator width is increased.

Closes #7